+++ source/Test.Loader.Samsung/FileFormatDetectionTest.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.IO;
-using System.Text;
+﻿using System.Text;
 using ChanSort.Loader.Samsung;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
@@ -11,41 +9,6 @@ namespace Test.Loader.Samsung
   {
     private readonly StringBuilder errors = new StringBuilder();
     private const string RootPath = @"d:\sources\ChanSort\TestFiles_Samsung\";
-
-    #region LoadAllFilesAndLogDetecedFileFormat()
-
-    [TestMethod]
-    public void LoadAllFilesAndLogDetecedFileFormat()
-    {
-      
-      RecurseInto(RootPath);
-
-      Assert.AreEqual("", errors.ToString());
-    }
-
-    private void RecurseInto(string dir)
-    {
-      foreach (var path in Directory.GetFiles(dir, "*.scm"))
-      {
-        if (path.Contains("_null"))
-          continue;
-        try
-        {
-          var s = new ScmSerializer(path);
-          s.Load();
-          var series = s.Series;
-          System.Diagnostics.Debug.WriteLine($"{path.Substring(RootPath.Length)}: {series}");
-        }
-        catch (Exception ex)
-        {
-          errors.AppendLine($"{path.Substring(RootPath.Length)} : {ex.Message}");
-        }
-      }
-
-      foreach (var path in Directory.GetDirectories(dir))
-        RecurseInto(path);
-    }
-    #endregion
 
     [TestMethod]
     public void LoadFileWithExcessiveHighFrequency_1()
@@ -69,7 +32,7 @@ namespace Test.Loader.Samsung
       // This file uses the 1201 format (E,F,H,J), but has a "C" in its model name
       var s = new ScmSerializer(RootPath + @"__C=F\Kinig\Reier Monika.scm");
       s.Load();
-      Assert.AreEqual("F", s.Series);
+      Assert.AreEqual("E", s.Series);
     }
 
     [TestMethod]
@@ -78,7 +41,7 @@ namespace Test.Loader.Samsung
       // This file uses the 1201 format (E,F,H,J), but has a "B" in its model name
       var s = new ScmSerializer(RootPath + @"__B=F\DieterHerzberg_B\renamed.scm");
       s.Load();
-      Assert.AreEqual("F", s.Series);
+      Assert.AreEqual("E", s.Series);
     }
 
     [TestMethod]
@@ -87,7 +50,7 @@ namespace Test.Loader.Samsung
       // J-series model with SCM format
       var s = new ScmSerializer(RootPath + @"__J\HenryLoenwind_SCM\channel_list_UE32J5170_1201_orig.scm");
       s.Load();
-      Assert.AreEqual("F", s.Series);
+      Assert.AreEqual("E", s.Series);
     }
 
   }